# ui/openapi/bizigo-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Bizigo.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/rca": {
      "post": {
        "tags": [
          "rca"
        ],
        "operationId": "GatherRca",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RcaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RcaReportResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RcaReportResponse"
                }
              }
            }
          },
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RcaRunResponse"
                }
              }
            }
          },
          "410": {
            "description": "Gone",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RcaAdmissionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "rca"
        ],
        "operationId": "ListRcaBundles",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RcaBundleListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rca/{id}": {
      "get": {
        "tags": [
          "rca"
        ],
        "operationId": "GetRcaReport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RcaReportResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "Gone",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rca/{id}/export": {
      "get": {
        "tags": [
          "rca"
        ],
        "operationId": "ExportRcaReport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "Gone",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rca/{id}/review": {
      "post": {
        "tags": [
          "rca"
        ],
        "operationId": "ReviewRcaReport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RcaReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RcaReviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "hint": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RcaAdmissionResponse": {
        "required": [
          "run_id",
          "accepted",
          "rejection",
          "detail"
        ],
        "type": "object",
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid"
          },
          "accepted": {
            "type": "boolean"
          },
          "rejection": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "RcaBundleListResponse": {
        "required": [
          "bundles"
        ],
        "type": "object",
        "properties": {
          "bundles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RcaBundleSummaryResponse"
            }
          }
        }
      },
...
      "RcaReportResponse": {
        "required": [
          "bundle_id",
          "content_hash",
          "gathered_at",
          "window",
          "findings",
          "timeline",
          "silent",
          "not_consulted",
          "trust",
          "out_of_scope_count",
          "is_partial",
          "review",
          "reasoning"
        ],
        "type": "object",
        "properties": {
          "bundle_id": {
            "type": "string",
            "format": "uuid"
          },
          "content_hash": {
            "type": "string"
          },
          "gathered_at": {
            "type": "string",
            "format": "date-time"
          },
          "window": {
            "$ref": "#/components/schemas/RcaWindowResponse"
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RcaFindingResponse"
            }
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RcaFindingResponse"
            }
          },
          "silent": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RcaSliceResponse"
            }
          },
          "not_consulted": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RcaSliceResponse"
            }
          },
          "trust": {
            "$ref": "#/components/schemas/RcaTrustResponse"
          },
          "out_of_scope_count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "is_partial": {
            "type": "boolean"
          },
          "review": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RcaReviewResponse"
              }
            ]
          },
          "reasoning": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RcaReasoningResponse"
              }
            ]
          },
          "excluded_input_records": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExcludedInputRecords"
              }
            ]
          },
          "coverage": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/EvidenceKindCoverage"
            }
          },
          "providers": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/RcaSliceResponse"
            }
          }
        }
      },
      "RcaRequest": {
        "required": [
          "from",
          "to",
          "baseline_from",
          "baseline_to"
        ],
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date-time"
          },
          "to": {
            "type": "string",
            "format": "date-time"
          },
          "baseline_from": {
            "type": "string",
            "format": "date-time"
          },
          "baseline_to": {
            "type": "string",
            "format": "date-time"
          },
          "owner_groups": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "source_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "RcaRunResponse": {
        "required": [
          "id",
          "source",
          "trigger_identity",
          "owner_group",
          "state",
          "reason",
          "counts_against_quota",
          "model_boundary",
          "model_boundary_override_reason",
          "requested_at",
          "finished_at"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "source": {
            "type": "string"
          },
          "trigger_identity": {
            "type": "string"
          },
          "owner_group": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "counts_against_quota": {
            "type": "boolean"
          },
          "model_boundary": {
            "type": "string"
          },
          "model_boundary_override_reason": {
            "type": [
              "null",
              "string"
            ]
          },
          "requested_at": {
            "type": "string",
            "format": "date-time"
          },
          "finished_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}